T8751 Improve Header codebase
-   Move out `<PromoLine />` to `<Layout />`;
-   Move out `alternativeHeader()` to `<Layout />`;
-   Remove `<React.Fragment />`.
