+++ src/components/Layout/Layout.js
@@ -6,13 +6,54 @@ import Footer from "../Footer";
 import ContactForm from "../ContactForm";
 import Constraint from "../Constraint";
 import Section from "../Section";
+import PromoLine from "../PromoLine";
+import Button from "../Button";
 
 import "./Layout.css";
 
 const Layout = ({ children, noStickyHeader, pagePath }) => {
+  const alternativeHeader = () => {
+    if (typeof window === `undefined`) {
+      return;
+    }
+    const pathName = window.location.pathname;
+    if (
+      pathName.startsWith(`/pagalba`) ||
+      pathName.startsWith(`/refugee-guide`)
+    ) {
+      return true;
+    }
+    return false;
+  };
+
   return (
     <div className="Layout">
-      <Header noSticky={noStickyHeader} />
+      <Header noSticky={noStickyHeader} altHeader={alternativeHeader()} />
+      {!alternativeHeader() && (
+        <PromoLine
+          title="ВСЯ ВАЖЛИВА ІНФОРМАЦІЯ ДЛЯ ГРОМАДЯН УКРАЇНИ"
+          titleLink="https://www.withukraine.lt"
+        >
+          <Button
+            icon={`arrow-blue`}
+            href="https://www.withukraine.lt"
+            color={`secondary`}
+            text={`ІНФОРМАЦІЯ`}
+            position={`right`}
+            target="_blank"
+            rel="noopener"
+          />
+          <Button
+            icon={`arrow-blue`}
+            href="https://www.withukraine.lt/help-search"
+            color={`secondary`}
+            text={`послуги`}
+            position={`right`}
+            target="_blank"
+            rel="noopener"
+          />
+        </PromoLine>
+      )}
       <main>{children}</main>
 
       <Section className="ContactFormSection" bgColor="blue">
